refactor: 리뷰 반영, 훅 분리
